Basic file compression implementation
Currently only works with CSS -- and only partially.

Working to develop a more robust Regex for maintaining necessary spaces
(such as those in the SVG data for the background).

Plan to implement smart compression for HTML and JavaScript as well.

TODO in future:
Figure out a more efficient way to handle server connections -- test if
it actually needs to be creating new threads, was leftover code from
original server creation.